+++ tests/src/sanity.rs
@@ -14,9 +14,14 @@ mod tests {
     use crate::utils::*;
 
     #[tokio::test]
+    async fn test_complete_node_api() {
+        _test_complete_node_api().await;
+    }
+
+    // actual test function used with "containerd", "docker" runtimes
     // starts the node(express.js) api server, sends four different requests, validates data,
     // stops the server and validates if the agent job and pod are deleted
-    async fn test_complete_node_api() {
+    async fn _test_complete_node_api() {
         let client = setup_kube_client().await;
 
         let pod_namespace = "default";
@@ -287,5 +292,12 @@ mod tests {
         .await
         .unwrap()
         .unwrap();
+    }
+
+    // docker runtime test
+    #[ignore]
+    #[tokio::test]
+    async fn test_docker_runtime() {
+        _test_complete_node_api().await;
     }
 }