+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## 2.1.0
 ### Added
 - Prompt user to update if their version is outdated in the VS Code extension or CLI.
-- Set `MIRRORD_CHECK_VERSION` to false to make E2E tests not read update messages. 
+- Add support for docker runtime, closes [#95](https://github.com/metalbear-co/mirrord/issues/95).
 - Add a keep-alive to keep the agent-pod from exiting, closes [#63](https://github.com/metalbear-co/mirrord/issues/63)
 
 ## 2.0.4